+++ core-bundle/tests/EventListener/BackendMenuListenerTest.php
@@ -45,18 +45,20 @@ class BackendMenuListenerTest extends TestCase
                     'label' => 'Category 1',
                     'title' => 'Category 1 Title',
                     'href' => '/',
-                    'class' => 'node-expanded',
+                    'class' => 'node-expanded trail custom-class',
                     'modules' => [
                         'node1' => [
                             'label' => 'Node 1',
                             'title' => 'Node 1 Title',
                             'href' => '/node1',
+                            'class' => 'node1',
                             'isActive' => true,
                         ],
                         'node2' => [
                             'label' => 'Node 2',
                             'title' => 'Node 2 Title',
                             'href' => '/node2',
+                            'class' => 'node2',
                             'isActive' => false,
                         ],
                     ],
@@ -102,6 +104,7 @@ class BackendMenuListenerTest extends TestCase
         $this->assertNotNull($categoryNode);
         $this->assertInstanceOf('Knp\Menu\ItemInterface', $categoryNode);
         $this->assertCount(2, $categoryNode->getChildren());
+        $this->assertSame('custom-class', $categoryNode->getAttribute('class'));
 
         // Test module node
         $moduleNode = $categoryNode->getChild('node1');